+++ test/e2e-test-application/e2e/tests/1-angular/user_settings_dialog.spec.js
@@ -327,7 +327,7 @@ describe('Navigation', () => {
       });
     });
 
-    it('Ctx update after storing user settings data using custom messages', () => {
+    it.only('Ctx update after storing user settings data using custom messages', () => {
       let $iframeBody;
       cy.get('[data-testid="us-navigation-item"]')
         .eq(5)
